Task system and builder rework.
