--- examples/hello_world/main.cpp
@@ -1,7 +1,0 @@
-#include <iostream>
-
-int main()
-{
-    std::cout << "Hello world!" << std::endl;
-    return 0;
-}
