+++ halfling/_runner.py
@@ -1,11 +1,12 @@
 """Task runner called by halfling CLI command."""
+
 import argparse
 import sys
 import importlib.util
 
-from halfling import _HALFLING_VERSION
 from halfling.exceptions import HalflingError
-from halfling.tasks.tasks import tasks # what a lovely heirarchy.......
+from halfling.tasks import tasks # what a lovely heirarchy.......
+from halfling.utils import _HALFLING_VERSION
 
 # TODO (aloebs29): search up directories for this file if it is not found in the current dir. This
 # would allow the halfling CLI command to be called anywhere within a project.
@@ -28,7 +29,8 @@ def _collect_command_line_args():
     subparsers = parser.add_subparsers()
     for name, task in tasks.items():
         task_parser = subparsers.add_parser(name)
-        task.add_args(task_parser)
+        if task.add_args is not None:
+            task.add_args(task_parser)
         task_parser.set_defaults(func=task.run)
 
     return parser.parse_args()
