+++ halfling/main.py
@@ -1,5 +1,6 @@
 """Main application called by halfling script."""
-from halfling.tasks._runner import run
+
+from halfling._runner import run
 
 
 def main():
